requirements.txt
clipboard
cryptography
ffmpy
m3u8
natsort
protobuf==3.20.*
pycaption
pycountry
pycryptodomex
pymediainfo
requests
svtplay_dl==3.7
titlecase
tqdm
Unidecode
pytest-playwright
